Move linkify.js to separate directory
